+++ NEWS.md
@@ -1,6 +1,6 @@
-# arkhe 1.5.0.9000
+# arkhe 1.6.0
 ## New classes and methods
-* Add `describe()` to quickly describe a table.
+* Add `describe()` to quickly describe a `matrix`-like object.
 * Add `sparsity()` to computes data sparsity (proportion of zeros).
 
 # arkhe 1.5.0